web,backend 3.0.0

+++ lite_agent_backend/lite-agent-rest/src/main/java/com/litevar/agent/rest/controller/v1/ModelController.java
@@ -1,7 +1,9 @@
 package com.litevar.agent.rest.controller.v1;
 
 import cn.hutool.core.lang.Validator;
+import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.StrUtil;
+import com.litevar.agent.auth.annotation.SystemRole;
 import com.litevar.agent.auth.annotation.WorkspaceRole;
 import com.litevar.agent.base.constant.CommonConstant;
 import com.litevar.agent.base.dto.ModelDTO;
@@ -10,10 +12,10 @@ import com.litevar.agent.base.entity.LlmModel;
 import com.litevar.agent.base.enums.AiProvider;
 import com.litevar.agent.base.enums.RoleEnum;
 import com.litevar.agent.base.enums.ServiceExceptionEnum;
+import com.litevar.agent.base.enums.SystemRoleEnum;
 import com.litevar.agent.base.exception.ServiceException;
 import com.litevar.agent.base.response.PageModel;
 import com.litevar.agent.base.response.ResponseData;
-import com.litevar.agent.base.util.LoginContext;
 import com.litevar.agent.base.valid.AddAction;
 import com.litevar.agent.base.valid.UpdateAction;
 import com.litevar.agent.base.vo.ModelVO;
@@ -62,6 +64,23 @@ public class ModelController {
     }
 
     /**
+     * 新建模型(系统管理员)
+     *
+     * @param modelVO
+     * @return
+     */
+    @PostMapping("/addForSystem")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData<String> model(@Validated(value = AddAction.class) @RequestBody ModelVO modelVO) {
+        boolean flag = Validator.isUrl(modelVO.getBaseUrl());
+        if (!flag) {
+            throw new ServiceException(ServiceExceptionEnum.ARGUMENT_NOT_VALID);
+        }
+        LlmModel model = modelService.addModel("0", modelVO);
+        return ResponseData.success(model.getId());
+    }
+
+    /**
      * 删除模型
      *
      * @param id 模型id
@@ -73,6 +92,28 @@ public class ModelController {
         List<Dataset> datasets = datasetService.searchDatasetsByLlmModelId(id);
         if (!datasets.isEmpty()) {
             return ResponseData.error(datasets);
+        }
+
+        modelService.removeModel(id);
+        return ResponseData.success();
+    }
+
+    /**
+     * 删除模型(系统管理员)
+     *
+     * @param id 模型id
+     * @return
+     */
+    @DeleteMapping("/deleteForSystem/{id}")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData deleteForSystem(@PathVariable("id") String id) {
+        List<Dataset> datasets = datasetService.searchDatasetsByLlmModelId(id);
+        if (!datasets.isEmpty()) {
+            return ResponseData.error(datasets);
+        }
+        LlmModel model = modelService.findById(id);
+        if (ObjectUtil.notEqual(model.getStatus(), 0)) {
+            throw new ServiceException(ServiceExceptionEnum.OPERATE_FAILURE);
         }
 
         modelService.removeModel(id);
@@ -96,21 +137,72 @@ public class ModelController {
     }
 
     /**
+     * 修改模型(系统管理员)
+     *
+     * @return
+     */
+    @PutMapping("/updateForSystem")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData<String> updateModelForSystem(@Validated(value = UpdateAction.class) @RequestBody ModelVO vo) {
+        boolean flag = !Validator.isUrl(vo.getBaseUrl()) || StrUtil.equals(vo.getApiKey(), "{{<APIKEY>}}");
+        if (flag) {
+            throw new ServiceException(ServiceExceptionEnum.ARGUMENT_NOT_VALID);
+        }
+        LlmModel model = modelService.findById(vo.getId());
+        if (ObjectUtil.equal(model.getStatus(), 1)) {
+            //已启用不能删除
+            throw new ServiceException(ServiceExceptionEnum.OPERATE_FAILURE);
+        }
+        modelService.updateModel(vo);
+        return ResponseData.success();
+    }
+
+    /**
      * 模型列表
      *
      * @param pageSize
      * @param pageNo
-     * @param type
+     * @param type      模型类型
+     * @param query     查询条件(支持查模型名字,别名)
+     * @param status    模型状态(1-启用, 2-禁用)
      * @param autoAgent 是否支持auto agent
+     * @param tab       0-全部,1-系统,2-我的
      * @return
      */
     @GetMapping("/list")
     public ResponseData<PageModel<ModelDTO>> model(@RequestHeader(CommonConstant.HEADER_WORKSPACE_ID) String workspaceId,
                                                    @RequestParam(value = "type", required = false) String type,
+                                                   @RequestParam(value = "query", required = false) String query,
+                                                   @RequestParam(value = "tab", required = false, defaultValue = "0") Integer tab,
+                                                   @RequestParam(value = "status", required = false) Integer status,
                                                    @RequestParam(value = "autoAgent", required = false) Boolean autoAgent,
                                                    @RequestParam(value = "pageSize", defaultValue = "10") Integer pageSize,
                                                    @RequestParam(value = "pageNo", defaultValue = "1") Integer pageNo) {
-        PageModel<ModelDTO> res = modelService.modelList(workspaceId, type, autoAgent, pageSize, pageNo);
+        //非系统用户只能看到已启用和停用的模型数据
+        List<Integer> statusList;
+        if (ObjectUtil.isNotEmpty(status)) {
+            statusList = List.of(status);
+        } else {
+            statusList = List.of(1, 2);
+        }
+        PageModel<ModelDTO> res = modelService.modelList(workspaceId, type, autoAgent, pageSize, pageNo, query, tab, statusList);
+        return ResponseData.success(res);
+    }
+
+    /**
+     * 模型列表(系统管理员)
+     *
+     * @param query    查询条件(支持查模型名字,别名)
+     * @param pageSize
+     * @param pageNo
+     * @return
+     */
+    @GetMapping("/listForSystem")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData<PageModel<ModelDTO>> model(@RequestParam(value = "query", required = false) String query,
+                                                   @RequestParam(value = "pageSize", defaultValue = "10") Integer pageSize,
+                                                   @RequestParam(value = "pageNo", defaultValue = "1") Integer pageNo) {
+        PageModel<ModelDTO> res = modelService.modelList("0", null, null, pageSize, pageNo, query, 1, null);
         return ResponseData.success(res);
     }
 
@@ -135,9 +227,6 @@ public class ModelController {
                             @PathVariable("id") String id,
                             @RequestParam(value = "plainText", defaultValue = "false") boolean plainText) {
         LlmModel model = modelService.findById(id);
-        if (!StrUtil.equals(LoginContext.currentUserId(), model.getUserId())) {
-            plainText = false;
-        }
         byte[] bytes = modelService.exportModel(model, plainText);
         FileDownloadUtil.download(response, model.getName() + ".json", bytes);
     }
@@ -155,5 +244,34 @@ public class ModelController {
                                                           @RequestParam("files") MultipartFile[] files) {
         Map<String, String> mapping = modelService.importModels(workspaceId, files);
         return ResponseData.success(mapping);
+    }
+
+    /**
+     * 导入模型配置(系统管理员)
+     *
+     * @param files 模型配置文件列表
+     * @return 导入结果
+     */
+    @PostMapping("/importForSystem")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData<Map<String, String>> importModels(@RequestParam("files") MultipartFile[] files) {
+        Map<String, String> mapping = modelService.importModels("0", files);
+        return ResponseData.success(mapping);
+    }
+
+    /**
+     * 切换模型状态(启用/禁用)
+     *
+     * @param id 模型ID
+     * @return
+     */
+    @PostMapping("/{id}/status/toggle")
+    @SystemRole(value = {SystemRoleEnum.ROLE_SYSTEM_ADMIN})
+    public ResponseData<String> toggleStatus(
+            @PathVariable("id") String id,
+            @RequestParam("status") Integer status
+    ) {
+        modelService.toggleStatus(id, status);
+        return ResponseData.success();
     }
 }
